Fix issue due to PlasmaCalendar returning wrong year
PlasmaCalendar year property is returning 0, what breaks the logic of
getting which was the day of the week of Jan 1st, current year.
This commit relies on JS Date object functions to get the current year.
Also, it fixes a old and dumb issue in this plasmoid due to the fact
January in JavaScript is 0, not 1.